# xiaoyang-web-http-pb-152026/api/xiaoyang/v1/team_router.pb.go
// Code generated by https://github.com/zhufuyi/sponge, DO NOT EDIT.

package v1

import (
	context "context"
	gin "github.com/gin-gonic/gin"
	errcode "github.com/zhufuyi/sponge/pkg/errcode"
	middleware "github.com/zhufuyi/sponge/pkg/gin/middleware"
	zap "go.uber.org/zap"
	strings "strings"
)

// import packages: strings. context. errcode. middleware. zap. gin.

type TeamServiceLogicer interface {
	CreateTeam(ctx context.Context, req *CreateTeamRequest) (*CreateTeamResponse, error)
	ListTeam(ctx context.Context, req *ListTeamRequest) (*ListTeamResponse, error)
	ListActiveTeam(ctx context.Context, req *ListTeamRequest) (*ListTeamResponse, error)
	UpdateTeam(ctx context.Context, req *UpdateTeamRequest) (*UpdateTeamResponse, error)
	SetTeamVisible(ctx context.Context, req *SetTeamVisibleRequest) (*SetTeamVisibleResponse, error)
	SetTeamLock(ctx context.Context, req *SetTeamVisibleRequest) (*SetTeamVisibleResponse, error)
	SetTeamClose(ctx context.Context, req *SetTeamVisibleRequest) (*SetTeamVisibleResponse, error)
	CreateTeamTemplate(ctx context.Context, req *CreateTeamTemplateRequest) (*CreateTeamTemplateResponse, error)
	ListTeamTemplate(ctx context.Context, req *ListTeamRequest) (*ListTeamResponse, error)
	DeleteTeamTemplate(ctx context.Context, req *SetTeamVisibleRequest) (*SetTeamVisibleResponse, error)
}

type TeamServiceOption func(*teamServiceOptions)

type teamServiceOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *teamServiceOptions) apply(opts ...TeamServiceOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithTeamServiceHTTPResponse() TeamServiceOption {
	return func(o *teamServiceOptions) {
		o.isFromRPC = false
	}
}

func WithTeamServiceRPCResponse() TeamServiceOption {
	return func(o *teamServiceOptions) {
		o.isFromRPC = true
	}
}

func WithTeamServiceResponser(responser errcode.Responser) TeamServiceOption {
	return func(o *teamServiceOptions) {
		o.responser = responser
	}
}

func WithTeamServiceLogger(zapLog *zap.Logger) TeamServiceOption {
	return func(o *teamServiceOptions) {
		o.zapLog = zapLog
	}
}

func WithTeamServiceErrorToHTTPCode(e ...*errcode.Error) TeamServiceOption {
	return func(o *teamServiceOptions) {
		o.httpErrors = e
	}
}

func WithTeamServiceRPCStatusToHTTPCode(s ...*errcode.RPCStatus) TeamServiceOption {
	return func(o *teamServiceOptions) {
		o.rpcStatus = s
	}
}

func WithTeamServiceWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) TeamServiceOption {
	return func(o *teamServiceOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterTeamServiceRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic TeamServiceLogicer,
	opts ...TeamServiceOption) {

	o := &teamServiceOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &teamServiceRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type teamServiceRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                TeamServiceLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *teamServiceRouter) register() {
	r.iRouter.Handle("POST", "/v1/team", r.withMiddleware("POST", "/v1/team", r.CreateTeam_0)...)
	r.iRouter.Handle("GET", "/v1/team/list", r.withMiddleware("GET", "/v1/team/list", r.ListTeam_0)...)
	r.iRouter.Handle("GET", "/v1/team/active", r.withMiddleware("GET", "/v1/team/active", r.ListActiveTeam_0)...)
	r.iRouter.Handle("PUT", "/v1/team", r.withMiddleware("PUT", "/v1/team", r.UpdateTeam_0)...)
	r.iRouter.Handle("POST", "/v1/team/visible", r.withMiddleware("POST", "/v1/team/visible", r.SetTeamVisible_0)...)
	r.iRouter.Handle("POST", "/v1/team/lock", r.withMiddleware("POST", "/v1/team/lock", r.SetTeamLock_0)...)
	r.iRouter.Handle("POST", "/v1/team/close", r.withMiddleware("POST", "/v1/team/close", r.SetTeamClose_0)...)
	r.iRouter.Handle("POST", "/v1/team_template", r.withMiddleware("POST", "/v1/team_template", r.CreateTeamTemplate_0)...)
	r.iRouter.Handle("GET", "/v1/team_template/list", r.withMiddleware("GET", "/v1/team_template/list", r.ListTeamTemplate_0)...)
	r.iRouter.Handle("DELETE", "/v1/team_template", r.withMiddleware("DELETE", "/v1/team_template", r.DeleteTeamTemplate_0)...)

}

func (r *teamServiceRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

var _ middleware.CtxKeyString

func (r *teamServiceRouter) CreateTeam_0(c *gin.Context) {
	req := &CreateTeamRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.CreateTeam(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) ListTeam_0(c *gin.Context) {
	req := &ListTeamRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListTeam(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) ListActiveTeam_0(c *gin.Context) {
	req := &ListTeamRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListActiveTeam(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) UpdateTeam_0(c *gin.Context) {
	req := &UpdateTeamRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateTeam(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) SetTeamVisible_0(c *gin.Context) {
	req := &SetTeamVisibleRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.SetTeamVisible(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) SetTeamLock_0(c *gin.Context) {
	req := &SetTeamVisibleRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.SetTeamLock(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) SetTeamClose_0(c *gin.Context) {
	req := &SetTeamVisibleRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.SetTeamClose(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) CreateTeamTemplate_0(c *gin.Context) {
	req := &CreateTeamTemplateRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.CreateTeamTemplate(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) ListTeamTemplate_0(c *gin.Context) {
	req := &ListTeamRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListTeamTemplate(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teamServiceRouter) DeleteTeamTemplate_0(c *gin.Context) {
	req := &SetTeamVisibleRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteTeamTemplate(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}