# xiaoyang-web-http-pb-152026/docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "xiaoyang api docs",
    "version": "2.0"
  },
...
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/team": {
      "post": {
        "summary": "开团",
        "description": "开团",
        "operationId": "TeamService_CreateTeam",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateTeamResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateTeamRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新开团信息",
        "description": "更新开团信息",
        "operationId": "TeamService_UpdateTeam",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateTeamResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateTeamRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/team/active": {
      "get": {
        "summary": "获取活跃的开团列表",
        "description": "获取活跃的开团列表",
        "operationId": "TeamService_ListActiveTeam",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListTeamResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "组织唯一标识",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "页码",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageSize",
            "description": "每页数量",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageCount",
            "description": "总页数",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/team/close": {
      "post": {
        "summary": "设定开团关闭状态",
        "description": "设定开团关闭状态",
        "operationId": "TeamService_SetTeamClose",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetTeamVisibleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SetTeamVisibleRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/team/list": {
      "get": {
        "summary": "获取开团列表",
        "description": "获取开团列表",
        "operationId": "TeamService_ListTeam",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListTeamResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "组织唯一标识",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "页码",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageSize",
            "description": "每页数量",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageCount",
            "description": "总页数",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/team/lock": {
      "post": {
        "summary": "设定开团锁定状态",
        "description": "设定开团锁定状态",
        "operationId": "TeamService_SetTeamLock",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetTeamVisibleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SetTeamVisibleRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/team/visible": {
      "post": {
        "summary": "设定开团可见性",
        "description": "设定开团可见性",
        "operationId": "TeamService_SetTeamVisible",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetTeamVisibleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SetTeamVisibleRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/team_template": {
      "delete": {
        "summary": "删除开团模板",
        "description": "删除开团模板",
        "operationId": "TeamService_DeleteTeamTemplate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetTeamVisibleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "teamId",
            "description": "开团id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "isVisible",
            "description": "是否对外可见",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "创建开团模板",
        "description": "创建开团模板",
        "operationId": "TeamService_CreateTeamTemplate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateTeamTemplateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateTeamTemplateRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/team_template/list": {
      "get": {
        "summary": "获取开团模板列表",
        "description": "获取开团模板列表",
        "operationId": "TeamService_ListTeamTemplate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListTeamResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "组织唯一标识",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "页码",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageSize",
            "description": "每页数量",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageCount",
            "description": "总页数",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "TeamService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CreateTeamRequest": {
      "type": "object",
      "properties": {
        "createrId": {
          "type": "integer",
          "format": "int64",
          "title": "创建者id"
        },
        "createrQqNumber": {
          "type": "string",
          "title": "创建者qq号"
        },
        "leagueId": {
          "type": "integer",
          "format": "int64",
          "title": "组织id"
        },
        "ukey": {
          "type": "string",
          "title": "组织唯一标识"
        },
        "title": {
          "type": "string",
          "title": "开团标题"
        },
        "teamTime": {
          "type": "integer",
          "format": "int64",
          "title": "开团时间"
        },
        "dungeonName": {
          "type": "string",
          "title": "副本名称"
        },
        "bookXuanjing": {
          "type": "boolean",
          "title": "是否有人预定玄晶"
        },
        "bookYuntie": {
          "type": "boolean",
          "title": "是否有人预定陨铁"
        },
        "isVisible": {
          "type": "boolean",
          "title": "是否对外可见"
        },
        "isLock": {
          "type": "boolean",
          "title": "是否锁定"
        },
        "rule": {
          "type": "string",
          "title": "规则"
        },
        "notice": {
          "type": "string",
          "title": "公告"
        }
      },
      "title": "开团请求"
    },
    "v1CreateTeamResponse": {
      "type": "object",
      "properties": {
        "teamId": {
          "type": "integer",
          "format": "int64",
          "title": "开团id"
        }
      },
      "title": "开团响应"
    },
    "v1CreateTeamTemplateRequest": {
      "type": "object",
      "properties": {
        "createrId": {
          "type": "integer",
          "format": "int64",
          "title": "创建者id"
        },
        "createrQqNumber": {
          "type": "string",
          "title": "创建者qq号"
        },
        "leagueId": {
          "type": "integer",
          "format": "int64",
          "title": "组织id"
        },
        "ukey": {
          "type": "string",
          "title": "组织唯一标识"
        },
        "title": {
          "type": "string",
          "title": "模板标题"
        },
        "rule": {
          "type": "string",
          "title": "规则"
        },
        "notice": {
          "type": "string",
          "title": "公告"
        }
      },
      "title": "创建开团模板请求"
    },
    "v1CreateTeamTemplateResponse": {
      "type": "object",
      "properties": {
        "templateId": {
          "type": "integer",
          "format": "int64",
          "title": "模板id"
        }
      },
      "title": "创建开团模板响应"
    },
...
    "v1ListTeamResponse": {
      "type": "object",
      "properties": {
        "teamInfos": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1TeamInfo"
          },
          "title": "开团信息"
        }
      },
      "title": "获取开团列表响应"
    },
...
    "v1SetTeamVisibleRequest": {
      "type": "object",
      "properties": {
        "teamId": {
          "type": "integer",
          "format": "int64",
          "title": "开团id"
        },
        "isVisible": {
          "type": "boolean",
          "title": "是否对外可见"
        }
      },
      "title": "设定开团可见性请求"
    },
    "v1SetTeamVisibleResponse": {
      "type": "object",
      "title": "设定开团可见性响应"
    },
...
    "v1TeamInfo": {
      "type": "object",
      "properties": {
        "teamId": {
          "type": "integer",
          "format": "int64",
          "title": "开团id"
        },
        "createrId": {
          "type": "integer",
          "format": "int64",
          "title": "创建者id"
        },
        "createrQqNumber": {
          "type": "string",
          "title": "创建者qq号"
        },
        "leagueId": {
          "type": "integer",
          "format": "int64",
          "title": "组织id"
        },
        "ukey": {
          "type": "string",
          "title": "组织唯一标识"
        },
        "title": {
          "type": "string",
          "title": "开团标题"
        },
        "teamTime": {
          "type": "integer",
          "format": "int64",
          "title": "开团时间"
        },
        "dungeonName": {
          "type": "string",
          "title": "副本名称"
        },
        "bookXuanjing": {
          "type": "boolean",
          "title": "是否有人预定玄晶"
        },
        "bookYuntie": {
          "type": "boolean",
          "title": "是否有人预定陨铁"
        },
        "isVisible": {
          "type": "boolean",
          "title": "是否对外可见"
        },
        "isLock": {
          "type": "boolean",
          "title": "是否锁定"
        },
        "rule": {
          "type": "string",
          "title": "规则"
        },
        "notice": {
          "type": "string",
          "title": "公告"
        },
        "createTime": {
          "type": "integer",
          "format": "int64",
          "title": "创建时间"
        },
        "updateTime": {
          "type": "integer",
          "format": "int64",
          "title": "更新时间"
        },
        "closeTime": {
          "type": "integer",
          "format": "int64",
          "title": "关闭时间"
        }
      },
      "title": "开团信息"
    },
...
    "v1UpdateTeamRequest": {
      "type": "object",
      "properties": {
        "teamId": {
          "type": "integer",
          "format": "int64",
          "title": "开团id"
        },
        "leagueId": {
          "type": "integer",
          "format": "int64",
          "title": "组织id"
        },
        "title": {
          "type": "string",
          "title": "开团标题"
        },
        "teamTime": {
          "type": "integer",
          "format": "int64",
          "title": "开团时间"
        },
        "dungeonName": {
          "type": "string",
          "title": "副本名称"
        },
        "bookXuanjing": {
          "type": "boolean",
          "title": "是否有人预定玄晶"
        },
        "bookYuntie": {
          "type": "boolean",
          "title": "是否有人预定陨铁"
        },
        "isVisible": {
          "type": "boolean",
          "title": "是否对外可见"
        },
        "isLock": {
          "type": "boolean",
          "title": "是否锁定"
        },
        "rule": {
          "type": "string",
          "title": "规则"
        },
        "notice": {
          "type": "string",
          "title": "公告"
        },
        "isClose": {
          "type": "boolean",
          "title": "是否关闭"
        }
      },
      "title": "更新开团信息请求"
    },
    "v1UpdateTeamResponse": {
      "type": "object",
      "title": "更新开团信息响应"
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Input a \"Bearer your-jwt-token\" to Value",
      "name": "Authorization",
      "in": "header"
    },
    "LeagueAuth": {
      "type": "apiKey",
      "description": "Input league key to Value Input league key to Value",
      "name": "LeagueKey",
      "in": "header"
    }
  }
}